Several pprof fixes
* fixing a sign overflow when packing SrcKeys to ints. Negative values
  in the lower 32 bits overwrote the high 32 bits with 0xFFFFFFFF.
* fixing a segfault/500 error in pprof/symbols when processing the last address.
  The code used std::stoll() and because the string was NOT 0-terminated,
  the following (garbage) character could randomly change the value:
  `0x1234+0x3456(7)` would produce "0x1234" (stopped at `+`) and
  `0x34567` because stoll just kept going. An attempt to decode Func/Unit from that
  value randomly caused a segfault or 500. Sometimes it just worked.
  Creating a temporary std::string instead.
  I searched folly/hhvm code for a stoll replacement that would take the string length,
  but haven't found anything.
* fixing a segfault in pprof/symbols when the input is empty (was reading NULL)
* fixing a segfault when enabling traces in cli (was reading NULL)

Reviewed By: @scannell

Differential Revision: D1254885

+++ hphp/runtime/vm/srckey.h
@@ -62,7 +62,7 @@ struct SrcKey : private boost::totally_ordered<SrcKey> {
   // Packed representation of SrcKeys for use in contexts where we
   // want atomicity.  (SrcDB.)
   AtomicInt toAtomicInt() const {
-    return uint64_t(getFuncId()) << 32 | uint64_t(offset());
+    return uint64_t(getFuncId()) << 32 | uint64_t(uint32_t(offset()));
   }
   static SrcKey fromAtomicInt(AtomicInt in) {
     return SrcKey { uint32_t(in >> 32), (Offset) int32_t(in & 0xffffffff) };
